#496 observable, refreshable AbstractNavigation (non-recursive)

+++ box/src/main/java/de/qabel/box/storage/command/DeleteFileChange.kt
@@ -4,14 +4,20 @@ import de.qabel.box.storage.*
 import de.qabel.box.storage.exceptions.QblStorageException
 import org.slf4j.LoggerFactory
 
-class DeleteFileChange(private val file: BoxFile, private val indexNavigation: IndexNavigation, private val writeBackend: StorageWriteBackend) : DirectoryMetadataChange<Unit>, Postprocessable {
+class DeleteFileChange(
+    val file: BoxFile,
+    private val indexNavigation: IndexNavigation,
+    private val writeBackend: StorageWriteBackend
+): DirectoryMetadataChange<Unit>, Postprocessable {
     private val logger by lazy { LoggerFactory.getLogger(DeleteFileChange::class.java) }
 
     override fun postprocess(dm: DirectoryMetadata, writeBackend: StorageWriteBackend)
         = writeBackend.deleteBlock(file.block)
 
     override fun execute(dm: DirectoryMetadata) {
-        dm.deleteFile(file)
+        if (dm.hasFile(file.name)) {
+            dm.deleteFile(file)
+        }
 
         if (file.isShared()) {
             removeSharesFromIndex()
